fix: quebras de layout login; altera diretorio do dashboard, para comportamento da sidebar;
