+++ src/app/app.component.html
@@ -1,5 +1,5 @@
 <div class="container">
-  <app-header *ngIf="showSidebarHeader"></app-header>
+  <!-- <app-header *ngIf="showSidebarHeader"></app-header> -->
   <div class="app-content" [class.side-bar-expanded]="isSideBarChecked">
     <app-side-bar *ngIf="showSidebarHeader"></app-side-bar>
     <div class="main-content">
