+++ src/app/app.component.scss
@@ -19,7 +19,7 @@
   display: flex;
   flex: 1;
   transition: margin-left 0.3s ease-in-out;
-  margin-top: 7rem;
+  // margin-top: 7rem;
   height: calc(100vh - 7rem);
 }
