+++ src/app/components/atendimento/atendimento.component.html
@@ -1,4 +1,4 @@
-<div class="flex min-h-full antialiased p-2 text-zinc-800">
+<div class="content flex h-full antialiased p-2 text-zinc-800">
   <div class="flex w-1/3 min-h-full antialiased p-4">
     <!-- Sidebar -->
     <div
