+++ src/app/components/atendimento/atendimento.component.scss
@@ -1,0 +1,4 @@
+.content {
+  margin-top: 7rem;
+  height: calc(100vh - 7rem);
+}
