+++ src/app/components/dashboard/dashboard.component.html
@@ -1,14 +1,18 @@
-<div class="dashboard">
+<div class="dashboard content">
   <div class="row">
     <div class="col-md-3 mb-4">
       <div class="card border-left-primary shadow h-100 py-2">
         <div class="card-body">
           <div class="row no-gutters align-items-center">
             <div class="col mr-2">
-              <div class="text-xs font-weight-bold text-primary text-uppercase mb-1">
+              <div
+                class="text-xs font-weight-bold text-primary text-uppercase mb-1"
+              >
                 Total Conversas Leads
               </div>
-              <div class="h5 mb-0 font-weight-bold text-gray-800">{{ totalConversasLeads }}</div>
+              <div class="h5 mb-0 font-weight-bold text-gray-800">
+                {{ totalConversasLeads }}
+              </div>
             </div>
             <div class="col-auto">
               <i class="fas fa-calendar fa-2x text-gray-300"></i>
@@ -22,10 +26,14 @@
         <div class="card-body">
           <div class="row no-gutters align-items-center">
             <div class="col mr-2">
-              <div class="text-xs font-weight-bold text-success text-uppercase mb-1">
+              <div
+                class="text-xs font-weight-bold text-success text-uppercase mb-1"
+              >
                 Total Conversas Alunos
               </div>
-              <div class="h5 mb-0 font-weight-bold text-gray-800">{{ totalConversasAlunos }}</div>
+              <div class="h5 mb-0 font-weight-bold text-gray-800">
+                {{ totalConversasAlunos }}
+              </div>
             </div>
             <div class="col-auto">
               <i class="fas fa-dollar-sign fa-2x text-gray-300"></i>
@@ -39,12 +47,16 @@
         <div class="card-body">
           <div class="row no-gutters align-items-center">
             <div class="col mr-2">
-              <div class="text-xs font-weight-bold text-info text-uppercase mb-1">
+              <div
+                class="text-xs font-weight-bold text-info text-uppercase mb-1"
+              >
                 Tasks
               </div>
               <div class="row no-gutters align-items-center">
                 <div class="col-auto">
-                  <div class="h5 mb-0 mr-3 font-weight-bold text-gray-800">50%</div>
+                  <div class="h5 mb-0 mr-3 font-weight-bold text-gray-800">
+                    50%
+                  </div>
                 </div>
                 <div class="col">
                   <div class="progress progress-sm mr-2">
@@ -72,7 +84,9 @@
         <div class="card-body">
           <div class="row no-gutters align-items-center">
             <div class="col mr-2">
-              <div class="text-xs font-weight-bold text-warning text-uppercase mb-1">
+              <div
+                class="text-xs font-weight-bold text-warning text-uppercase mb-1"
+              >
                 Pending Requests
               </div>
               <div class="h5 mb-0 font-weight-bold text-gray-800">18</div>
@@ -90,34 +104,42 @@
     <div class="col-lg-8 mb-4">
       <div class="card shadow mb-4">
         <div class="card-header py-3">
-          <h6 class="m-0 font-weight-bold text-primary">Perguntas Mais Frequentes</h6>
+          <h6 class="m-0 font-weight-bold text-primary">
+            Perguntas Mais Frequentes
+          </h6>
         </div>
         <div class="card-body">
           <table class="table table-bordered">
             <thead>
-            <tr>
-              <th scope="col">Pergunta</th>
-              <th scope="col">Quantidade</th>
-            </tr>
+              <tr>
+                <th scope="col">Pergunta</th>
+                <th scope="col">Quantidade</th>
+              </tr>
             </thead>
             <tbody>
-            <tr *ngFor="let pergunta of perguntasFrequentesArray">
-              <td>{{ pergunta.pergunta }}</td>
-              <td>{{ pergunta.quantidade }}</td>
-            </tr>
+              <tr *ngFor="let pergunta of perguntasFrequentesArray">
+                <td>{{ pergunta.pergunta }}</td>
+                <td>{{ pergunta.quantidade }}</td>
+              </tr>
             </tbody>
-
           </table>
           <nav>
             <ul class="pagination justify-content-end">
               <li class="page-item" [class.disabled]="!previousPage">
-                <a class="page-link" href="/relatorios" (click)="changePage(previousPage)">Anterior</a>
+                <a
+                  class="page-link"
+                  href="/relatorios"
+                  (click)="changePage(previousPage)"
+                  >Anterior</a
+                >
               </li>
               <li class="page-item disabled">
-                <span class="page-link">Página {{ currentPage + 1 }} de {{ totalPages }}</span>
+                <span class="page-link"
+                  >Página {{ currentPage + 1 }} de {{ totalPages }}</span
+                >
               </li>
               <li class="page-item" [class.disabled]="!nextPage">
-                <a class="page-link"  (click)="changePage(nextPage)">Próxima</a>
+                <a class="page-link" (click)="changePage(nextPage)">Próxima</a>
               </li>
             </ul>
           </nav>
@@ -127,7 +149,9 @@
     <div class="col-lg-4 mb-4">
       <div class="card shadow mb-4">
         <div class="card-header py-3">
-          <h6 class="m-0 font-weight-bold text-primary">Total de Conversas Mês</h6>
+          <h6 class="m-0 font-weight-bold text-primary">
+            Total de Conversas Mês
+          </h6>
         </div>
         <div class="card-body">
           <canvas id="myPieChart"></canvas>
