+++ src/app/components/historico/historico.component.html
@@ -19,7 +19,7 @@
       </div>
       <div class="hidden shrink-0 sm:flex sm:flex-col sm:items-end">
         <button
-          class="bg-purple-500 rounded-lg text-zinc-50 text-semibold text-xl py-3 px-4"
+          class="bg-purple-600 rounded-lg text-zinc-50 text-semibold text-xl py-3 px-4 shadow-sm"
           (click)="verConversa(conversa.sessionId)"
         >
           Ver Conversa
