+++ src/app/components/login/login.component.html
@@ -1,12 +1,16 @@
 <div class="w-full h-full p-0 m-0 flex justify-center items-center">
   <div
-    class="card-login flex flex-col lg:flex-row items-center justify-center w-4/5 h-2/3 lg:w-9/12 lg:h-3/5 rounded-lg shadow space-y-6 lg:space-y-0 lg:space-x-6 p-6 lg:p-12"
+    class="card-login flex flex-col items-center justify-center w-5/6 h-5/6 rounded-lg shadow space-y-6 lg:flex-row lg:w-9/12 lg:h-3/5 lg:space-y-0 lg:space-x-6 p-6 lg:p-12"
   >
-    <form class="space-y-4 lg:w-2/5 lg:px-3" action="#" method="POST">
+    <form
+      class="space-y-6 w-5/6 md:w-3/4 lg:w-2/5 lg:px-3"
+      action="#"
+      method="POST"
+    >
       <div>
         <label
           for="login"
-          class="block text-lg font-medium leading-6 text-gray-900"
+          class="block text-2xl pb-1 font-sans font-medium text-blue-gray-900"
           >Login</label
         >
         <div class="mt-2">
@@ -17,43 +21,43 @@
             type="login"
             autocomplete="login"
             required
-            class="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-purple-600 sm:text-lg sm:leading-6"
+            class="block w-full rounded-lg py-3 px-4 text-2xl text-zinc-800 shadow-sm ring-1 focus:ring-2 focus:ring-inset focus:ring-purple-600"
           />
         </div>
       </div>
 
       <div>
-        <div class="flex items-center justify-between">
-          <label
-            for="password"
-            class="block text-lg font-medium leading-6 text-gray-900"
-            >Senha</label
-          >
-        </div>
-        <div class="mt-2">
-          <input
-            [(ngModel)]="usuario.senha"
-            id="password"
-            name="password"
-            type="password"
-            autocomplete="current-password"
-            required
-            class="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-purple-600 sm:text-lg sm:leading-6"
-          />
-        </div>
+        <label
+          for="password"
+          class="block text-2xl pb-1 font-sans font-medium text-blue-gray-900"
+          >Senha</label
+        >
+      </div>
+      <div class="mt-2">
+        <input
+          [(ngModel)]="usuario.senha"
+          id="password"
+          name="password"
+          type="password"
+          autocomplete="current-password"
+          required
+          class="block w-full rounded-lg py-3 px-4 text-2xl text-zinc-800 shadow-sm ring-1 focus:ring-2 focus:ring-inset focus:ring-purple-600 border-0"
+        />
       </div>
 
       <div>
         <button
           (click)="login()"
           type="submit"
-          class="flex w-full justify-center rounded-md bg-purple-600 px-3 py-1.5 text-lg font-semibold leading-6 text-white shadow-sm hover:bg-purple-500 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-purple-600"
+          class="flex w-full justify-center rounded-lg bg-purple-600 px-4 py-3 text-2xl font-sans font-medium text-zinc-50 shadow-sm hover:bg-purple-500 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-purple-600"
         >
           Entrar
         </button>
       </div>
-      <div class="text-lg text-center">
-        <a href="#" class="font-semibold text-purple-600 hover:text-purple-500"
+      <div class="text-2xl text-center md:w-3/4">
+        <a
+          href="#"
+          class="font-sans font-medium text-purple-600 hover:text-purple-500"
           >Esqueceu sua senha?</a
         >
       </div>
@@ -61,14 +65,9 @@
 
     <div class="min-h-custom p-0 m-0 flex justify-center items-center">
       <div
-        class="flex flex-col lg:flex-row items-center justify-center w-4/5 h-2/3 p-20"
+        class="flex flex-col lg:flex-row items-center justify-center w-4/5 h-4/5 m-1 md:w-3/5:h-3/5 pt-12"
       >
-        <a
-          routerLink="/"
-          class="logo py-2 lg:h-3/6 lg:px-4"
-          height="250"
-          width="250"
-        >
+        <a routerLink="/" class="logo py-2 lg:h-3/6 lg:px-4">
           <img src="../../../assets/images/Group1.svg" alt="Logo" />
         </a>
       </div>
