+++ src/app/components/login/login.component.scss
@@ -1,14 +1,3 @@
-// .container {
-//   display: flex;
-//   align-items: center;
-//   justify-content: center;
-//   min-height: calc(100vh - 10rem);
-// }
-
 .card-login {
   background: rgba(154, 123, 179, 0.5);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  box-shadow: 0 2rem 5rem rgba(0, 0, 0, 0.1);
 }
